[Pipelines] Move get_workflow_steps to helpers (#1 send notifications from BE)
